Added support for getting items from safe_array objects using accessor methods

+++ src/math_util.h
@@ -34,13 +34,13 @@
   #include <Eigen/Dense>
   using namespace Eigen;
 
-  typedef Matrix<int, Eigen::Dynamic, 1>  int_array;
+  typedef Matrix<int32_t, Eigen::Dynamic, 1>  int_array;
   typedef Matrix<nec_float, Eigen::Dynamic, 1>  real_array;
   typedef Matrix<nec_complex, Eigen::Dynamic, 1>  complex_array;
 #else
   // Use our own types rather than Eigen
   #include "safe_array.h"
-  typedef safe_array<int>  int_array;
+  typedef safe_array<int32_t>  int_array;
   typedef safe_array<nec_float>  real_array;
   typedef safe_array<nec_complex>  complex_array;
 
@@ -49,7 +49,7 @@
 #endif
 
 
-inline void vector_fill(complex_array& x, long start, long N, const nec_complex& y) {
+inline void vector_fill(complex_array& x, int32_t start, int32_t N, const nec_complex& y) {
   x.fill(start, N, y);
 }
